+++ MoreShipUpgrades/UpgradeComponents/Items/Contracts/Exorcism/PentagramScript.cs
@@ -1,6 +1,7 @@
 ﻿using GameNetcodeStuff;
 using MoreShipUpgrades.Managers;
 using MoreShipUpgrades.Misc;
+using MoreShipUpgrades.Misc.Util;
 using MoreShipUpgrades.UpgradeComponents.Contracts;
 using System.Collections;
 using System.Collections.Generic;